+++ crates/bevy_reactor_core/src/hover.rs
@@ -7,28 +7,6 @@ use crate::{signal::Signal, Cx, RunContextRead, RunContextSetup};
 #[derive(Default, Component)]
 pub struct Hovering(pub bool);
 
-// Note: previously this was implemented as a Reaction, however it was reacting every frame
-// because HoverMap is mutated every frame regardless of whether or not it changed.
-pub fn update_hover_states(
-    hover_map: Option<Res<HoverMap>>,
-    mut hovers: Query<(Entity, &mut Hovering)>,
-    parent_query: Query<&Parent>,
-) {
-    let Some(hover_map) = hover_map else { return };
-    let hover_set = hover_map.get(&PointerId::Mouse);
-    for (entity, mut hoverable) in hovers.iter_mut() {
-        let is_hovering = match hover_set {
-            Some(map) => map
-                .iter()
-                .any(|(ha, _)| parent_query.iter_ancestors(*ha).any(|e| e == entity)),
-            None => false,
-        };
-        if hoverable.0 != is_hovering {
-            hoverable.0 = is_hovering;
-        }
-    }
-}
-
 /// Method to create a signal that tracks whether the mouse is hovering over the given entity.
 pub trait CreateHoverSignal {
     /// Signal that returns true when the mouse is hovering over the given entity or a descendant.
