+++ crates/bevy_reactor_core/src/lib.rs
@@ -1,6 +1,5 @@
 // mod style;
 mod reaction;
-mod reactive_view;
 mod tracking_scope;
 mod derived;
 mod mutable;
@@ -26,7 +25,6 @@ pub use cx::RunContextWrite;
 pub use derived::Derived;
 pub use derived::ReadDerived;
 
-
 pub use reaction::*;
 pub use signal::IntoSignal;
 pub use signal::Signal;
@@ -40,9 +38,7 @@ pub use effect_target::EntityEffect;
 pub use tracking_scope::DespawnScopes;
 pub use tracking_scope::TrackingScope;
 pub use tracking_scope::TrackingScopeTracing;
-pub use tracking_scope::run_reactions;
 
 pub use node_span::NodeSpan;
-pub use hover::CreateHoverSignal;
-pub use hover::update_hover_states;
+pub use hover::{CreateHoverSignal, Hovering};
