+++ crates/bevy_reactor_plugin/Cargo.toml
@@ -5,6 +5,7 @@ edition = "2021"
 
 [dependencies]
 bevy = {workspace=true}
+bevy_mod_picking = { workspace=true }
 bevy_reactor_core = {path ="../bevy_reactor_core"}
 bevy_reactor_style = {path ="../bevy_reactor_style"}
 bevy_reactor_view = {path ="../bevy_reactor_view"}
