+++ crates/bevy_reactor_plugin/src/lib.rs
@@ -1,11 +1,9 @@
+mod systems;
+
 use bevy::prelude::*;
 
-use bevy_reactor_core::{
-    update_hover_states, run_reactions
-};
 
-use bevy_reactor_view::{attach_child_views, build_added_view_roots};
-use bevy_reactor_view::{update_text_styles, update_compositor_size,};
+use crate::systems::*;
 
 /// Plugin that adds the reactive UI system to the app.
 pub struct ReactorPlugin;
